Rename poses and put them in order
