+++ src/routes/poses/6_UpwardFacingDog.svelte
